src/test/java/dal/asd/dpl/InternalStateMachineTest/AdvanceTimeStateTest.java
package dal.asd.dpl.InternalStateMachineTest;

import dal.asd.dpl.InternalStateMachine.AdvanceTimeState;
import dal.asd.dpl.InternalStateMachine.InternalStateContext;
import dal.asd.dpl.UserInput.CmdUserInput;
import dal.asd.dpl.UserInput.IUserInput;
import dal.asd.dpl.UserOutput.CmdUserOutput;
import dal.asd.dpl.UserOutput.IUserOutput;
import org.junit.Before;
import org.junit.Test;

import java.util.Calendar;

import static org.junit.Assert.*;

public class AdvanceTimeStateTest {
    private static AdvanceTimeState state;
    private static IUserInput input;
    private static IUserOutput output;
    private static InternalStateContext context;
    private static Calendar seasonCalendar;

    @Before
    public void setUp() throws Exception {
        input = new CmdUserInput();
        output = new CmdUserOutput();
        seasonCalendar = Calendar.getInstance();
        state = new AdvanceTimeState(seasonCalendar,"02-04-2021", "03-04-2021", input, output);
        context = new InternalStateContext(input, output);
    }

    @Test
    public void nextStateTest() {
        context.setState(state);
        context.nextState();
        assertEquals("Training", state.getNextStateName());
    }

    @Test
    public void doProcessingTest() {
        state.doProcessing();
        assertEquals("03-04-2021", state.getCurrentDate());
    }

    @Test
    public void getStateNameTest() {
        assertEquals("AdvanceTime", state.getStateName());
    }

    @Test
    public void getCurrentDateTest() {
        assertEquals("02-04-2021", state.getCurrentDate());
    }

    @Test
    public void getNextStateNameTest() {
        state.nextState(context);
        assertEquals("Training", state.getNextStateName());
    }
}